+++ CHANGELOG.md
@@ -52,6 +52,7 @@ Version 1.4.2
 - fix infinite loop when failed to load Bink video (RF bug)
 - fix memory leak when trying to load non-existing Bink video (RF bug)
 - fix crash caused by explosion near dying player-controlled entity (RF bug)
+- fix unwanted fire after weapon switch from weapon with impact delay (RF bug)
 
 Version 1.4.1 (released 2019-12-07)
 -----------------------------------